requirements.txt
langchain
langchain_core
langchain_community
streamlit
python-dotenv
langchain-openai
langsmith
langchain_huggingface
langchain_groq
arxiv
wikipedia
duckduckgo-search
mysql-connector-python
